# internal/server/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package gen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Generate vector in embedding model
	// (POST /generate)
	GenerateVectors(ctx echo.Context) error
	// Get active embedding models
	// (GET /models)
	GetModels(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GenerateVectors converts echo context to params.
func (w *ServerInterfaceWrapper) GenerateVectors(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GenerateVectors(ctx)
	return err
}

// GetModels converts echo context to params.
func (w *ServerInterfaceWrapper) GetModels(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetModels(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/generate", wrapper.GenerateVectors)
	router.GET(baseURL+"/models", wrapper.GetModels)

}

type N400JSONResponse Error

type N404JSONResponse Error

type N500JSONResponse Error

type GenerateVectorsRequestObject struct {
	Body *GenerateVectorsJSONRequestBody
}

type GenerateVectorsResponseObject interface {
	VisitGenerateVectorsResponse(w http.ResponseWriter) error
}

type GenerateVectors200JSONResponse GeneratedEmbeddings

func (response GenerateVectors200JSONResponse) VisitGenerateVectorsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GenerateVectors404JSONResponse struct{ N404JSONResponse }

func (response GenerateVectors404JSONResponse) VisitGenerateVectorsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type GenerateVectors500JSONResponse struct{ N500JSONResponse }

func (response GenerateVectors500JSONResponse) VisitGenerateVectorsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetModelsRequestObject struct {
}

type GetModelsResponseObject interface {
	VisitGetModelsResponse(w http.ResponseWriter) error
}

type GetModels200JSONResponse Models

func (response GetModels200JSONResponse) VisitGetModelsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetModels400JSONResponse struct{ N400JSONResponse }

func (response GetModels400JSONResponse) VisitGetModelsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetModels500JSONResponse struct{ N500JSONResponse }

func (response GetModels500JSONResponse) VisitGetModelsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Generate vector in embedding model
	// (POST /generate)
	GenerateVectors(ctx context.Context, request GenerateVectorsRequestObject) (GenerateVectorsResponseObject, error)
	// Get active embedding models
	// (GET /models)
	GetModels(ctx context.Context, request GetModelsRequestObject) (GetModelsResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GenerateVectors operation middleware
func (sh *strictHandler) GenerateVectors(ctx echo.Context) error {
	var request GenerateVectorsRequestObject

	var body GenerateVectorsJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GenerateVectors(ctx.Request().Context(), request.(GenerateVectorsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GenerateVectors")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GenerateVectorsResponseObject); ok {
		return validResponse.VisitGenerateVectorsResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetModels operation middleware
func (sh *strictHandler) GetModels(ctx echo.Context) error {
	var request GetModelsRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetModels(ctx.Request().Context(), request.(GetModelsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetModels")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetModelsResponseObject); ok {
		return validResponse.VisitGetModelsResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}