# internal/server/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package gen

// Defines values for GenerateEmbeddingsInputType.
const (
	Document GenerateEmbeddingsInputType = "document"
	Query    GenerateEmbeddingsInputType = "query"
)

// EmbeddingData defines model for EmbeddingData.
type EmbeddingData struct {
	// Id Input text correlation id
	Id string `json:"id"`

	// Vector String to embed
	Vector []float32 `json:"vector"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// GenerateEmbeddings defines model for GenerateEmbeddings.
type GenerateEmbeddings struct {
	Input []InputData `json:"input"`

	// InputType Type of text you embedding, query (question) or document (preparing knowledge base)
	InputType GenerateEmbeddingsInputType `json:"input_type"`

	// Model Embedding model name
	Model string `json:"model"`
}

// GenerateEmbeddingsInputType Type of text you embedding, query (question) or document (preparing knowledge base)
type GenerateEmbeddingsInputType string

// GeneratedEmbeddings defines model for GeneratedEmbeddings.
type GeneratedEmbeddings struct {
	// Dimensions Resulting vector length
	Dimensions int             `json:"dimensions"`
	Embeddings []EmbeddingData `json:"embeddings"`

	// Model Embedding model name
	Model string `json:"model"`
}

// InputData defines model for InputData.
type InputData struct {
	// Id Input text correlation id
	Id string `json:"id"`

	// Text String to embed
	Text string `json:"text"`
}

// Model defines model for Model.
type Model struct {
	// Dimensions Resulting vector length
	Dimensions int `json:"dimensions"`

	// MaxInputTokens Maximum input tokens that model can operate
	MaxInputTokens int `json:"max_input_tokens"`

	// Multilingual Can this model operate with non-english languages
	Multilingual bool `json:"multilingual"`

	// Name Embedding model name
	Name string `json:"name"`
}

// Models defines model for Models.
type Models struct {
	Models []Model `json:"models"`
}

// N400 defines model for 400.
type N400 = Error

// N404 defines model for 404.
type N404 = Error

// N500 defines model for 500.
type N500 = Error

// GenerateVectorsJSONRequestBody defines body for GenerateVectors for application/json ContentType.
type GenerateVectorsJSONRequestBody = GenerateEmbeddings